enhance(unified):
Unified into a single model for convenience
