+++ pyproject.toml
@@ -3,13 +3,14 @@ name = "data-complexity"
 description = "Data Complexity Measures"
 authors = [
     {name = "Son Nguyen", email = "ngocson2vn@gmail.com"},
+    {name = "Iraitz Montalban", email = "iraitzm@gmail.com"}
 ]
 dependencies = [
     "numpy>=1.25.2",
     "gower>=0.1.2",
     "networkx>=3.2.1",
 ]
-requires-python = ">=3.9, <3.14"
+requires-python = ">=3.9, <3.11"
 readme = "README.md"
 license = {text = "MIT"}
 dynamic = ["version"]
@@ -23,7 +24,7 @@ includes = ["dcm/"]
 
 [tool.pdm.version]
 source = "file"
-path = "dcm/__init__.py"
+path = "dcm/__version__.py"
 
 [tool.pdm.dev-dependencies]
 examples = [
@@ -33,6 +34,8 @@ examples = [
     "scikit-learn>=1.3.1",
     "ipykernel>=6.25.2",
     "seaborn>=0.13.0",
+    "ydata-profiling>=4.10.0",
+    "setuptools>=75.1.0",
 ]
 tests = [
     "nose2[coverage_plugin]>=0.15.1",
