+++ tests/test_balance.py
@@ -4,7 +4,7 @@ Filename: test_balance.py
 
 import numpy
 import unittest
-from dcm import BalanceMeasures
+from dcm import ComplexityProfile
 from sklearn.datasets import load_iris, load_breast_cancer
 
 
@@ -16,14 +16,14 @@ class TestBalance(unittest.TestCase):
         self.y = y
 
     def test_iris(self):
-        model = BalanceMeasures()
+        model = ComplexityProfile(measures=["B1", "B2"])
         model.fit(self.X, self.y)
 
         numpy.testing.assert_allclose(model.B1(), 2.220446e-16, atol=0)
         numpy.testing.assert_allclose(model.B2(), 0.0, atol=0)
 
     def test_iris_100(self):
-        model = BalanceMeasures()
+        model = ComplexityProfile(measures=["B1", "B2"])
         model.fit(self.X[:100], self.y[:100])
 
         numpy.testing.assert_allclose(model.B1(), 0.0, atol=0)
@@ -38,14 +38,14 @@ class TestBalance2(unittest.TestCase):
         self.y = y
 
     def test_cancer(self):
-        model = BalanceMeasures()
+        model = ComplexityProfile(measures=["B1", "B2"])
         model.fit(self.X, self.y)
 
         numpy.testing.assert_allclose(model.B1(), 0.047365, atol=1e-4)
         numpy.testing.assert_allclose(model.B2(), 0.12196, atol=1e-4)
 
     def test_cancer_100(self):
-        model = BalanceMeasures()
+        model = ComplexityProfile(measures=["B1", "B2"])
         model.fit(self.X[:100], self.y[:100])
 
         numpy.testing.assert_allclose(model.B1(), 0.07, atol=1e-2)
