+++ tests/test_imbalance.py
@@ -4,7 +4,7 @@ Filename: test_imbalance.py
 
 import numpy
 import unittest
-from dcm import ImbalanceMeasures
+from dcm import ComplexityProfile
 from sklearn.datasets import load_iris, load_breast_cancer
 
 
@@ -16,14 +16,14 @@ class TestImbalance(unittest.TestCase):
         self.y = y
 
     def test_iris(self):
-        model = ImbalanceMeasures()
+        model = ComplexityProfile()
         model.fit(self.X, self.y)
 
         numpy.testing.assert_allclose(model.C1(), 1.0, atol=1e-06)
         numpy.testing.assert_allclose(model.C2(), 0.0, atol=1e-06)
 
     def test_iris_100(self):
-        model = ImbalanceMeasures()
+        model = ComplexityProfile()
         model.fit(self.X[:100], self.y[:100])
 
         numpy.testing.assert_allclose(model.C1(), 1.0, atol=1e-06)
@@ -38,14 +38,14 @@ class TestImbalance2(unittest.TestCase):
         self.y = y
 
     def test_cancer(self):
-        model = ImbalanceMeasures()
+        model = ComplexityProfile()
         model.fit(self.X, self.y)
 
         numpy.testing.assert_allclose(model.C1(), 0.952635, atol=1e-06)
         numpy.testing.assert_allclose(model.C2(), 0.12196, atol=1e-06)
 
     def test_cancer_100(self):
-        model = ImbalanceMeasures()
+        model = ComplexityProfile()
         model.fit(self.X[:100], self.y[:100])
 
         numpy.testing.assert_allclose(model.C1(), 0.934068, atol=1e-06)
